Restructure to perform full upload via script
This is a breaking change: the org_json_path field has been removed from the
org-toggl config file, and `org-toggl.py` now takes the JSON path as an
argument.

Introduce `org-toggl.sh`, a script which executes the full processing / upload
workflow: it runs Emacs in batch mode to export the given org file as JSON, and
then runs the Python script to upload it via the Toggl API. These were
previously separate steps - you had to export to JSON from within Emacs and
use the Python script to upload to Toggl.